requirements.txt
scipy
matplotlib
numpy
Pillow